refactor: optimize BFD traffic test (#15550)
Description of PR
Optimize the BFD traffic test to reduce the running time.

Summary:
Fixes # (issue) Microsoft ADO 30056122

Approach
What is the motivation for this PR?
There are unnecessary setup steps in the BFD traffic test, which can be removed to reduce the running time. The running time will be decreased by at least 35 min after this change.

How did you do it?
How did you verify/test it?
I ran the updated code and can confirm it's working well.

co-authorized by: jianquanye@microsoft.com
