+++ tests/bfd/test_bfd_traffic.py
@@ -1,11 +1,12 @@
 import logging
+import random
 
 import pytest
 
-from tests.bfd.bfd_base import BfdBase
 from tests.bfd.bfd_helpers import get_ptf_src_port, get_backend_interface_in_use_by_counter, \
     get_random_bgp_neighbor_ip_of_asic, toggle_port_channel_or_member, get_port_channel_by_member, \
-    wait_until_given_bfd_down, assert_traffic_switching, create_and_verify_bfd_state, verify_bfd_only
+    wait_until_given_bfd_down, assert_traffic_switching, verify_bfd_only, extract_backend_portchannels, \
+    get_src_dst_asic_next_hops
 from tests.common.helpers.multi_thread_utils import SafeThreadPoolExecutor
 
 pytestmark = [
@@ -16,8 +17,98 @@ pytestmark = [
 logger = logging.getLogger(__name__)
 
 
-class TestBfdTraffic(BfdBase):
+class TestBfdTraffic:
     PACKET_COUNT = 10000
+
+    @pytest.fixture(scope="class")
+    def select_dut_and_src_dst_asic_index(self, duthosts):
+        if not duthosts.frontend_nodes:
+            pytest.skip("DUT does not have any frontend nodes")
+
+        dut_index = random.choice(list(range(len(duthosts.frontend_nodes))))
+        asic_namespace_list = duthosts.frontend_nodes[dut_index].get_asic_namespace_list()
+        if len(asic_namespace_list) < 2:
+            pytest.skip("DUT does not have more than one ASICs")
+
+        # Random selection of src asic & dst asic on DUT
+        src_asic_namespace, dst_asic_namespace = random.sample(asic_namespace_list, 2)
+        src_asic_index = src_asic_namespace.split("asic")[1]
+        dst_asic_index = dst_asic_namespace.split("asic")[1]
+
+        yield {
+            "dut_index": dut_index,
+            "src_asic_index": int(src_asic_index),
+            "dst_asic_index": int(dst_asic_index),
+        }
+
+    @pytest.fixture(scope="class")
+    def get_src_dst_asic(self, request, duthosts, select_dut_and_src_dst_asic_index):
+        logger.info("Printing select_dut_and_src_dst_asic_index")
+        logger.info(select_dut_and_src_dst_asic_index)
+
+        logger.info("Printing duthosts.frontend_nodes")
+        logger.info(duthosts.frontend_nodes)
+        dut = duthosts.frontend_nodes[select_dut_and_src_dst_asic_index["dut_index"]]
+
+        logger.info("Printing dut asics")
+        logger.info(dut.asics)
+
+        src_asic = dut.asics[select_dut_and_src_dst_asic_index["src_asic_index"]]
+        dst_asic = dut.asics[select_dut_and_src_dst_asic_index["dst_asic_index"]]
+
+        request.config.src_asic = src_asic
+        request.config.dst_asic = dst_asic
+        request.config.dut = dut
+
+        rtn_dict = {
+            "src_asic": src_asic,
+            "dst_asic": dst_asic,
+            "dut": dut,
+        }
+
+        rtn_dict.update(select_dut_and_src_dst_asic_index)
+        yield rtn_dict
+
+    @pytest.fixture(scope="class", params=["ipv4", "ipv6"])
+    def prepare_traffic_test_variables(self, get_src_dst_asic, request):
+        version = request.param
+        logger.info("Version: %s", version)
+
+        dut = get_src_dst_asic["dut"]
+        src_asic = get_src_dst_asic["src_asic"]
+        src_asic_index = get_src_dst_asic["src_asic_index"]
+        dst_asic = get_src_dst_asic["dst_asic"]
+        dst_asic_index = get_src_dst_asic["dst_asic_index"]
+        logger.info(
+            "DUT: {}, src_asic_index: {}, dst_asic_index: {}".format(dut.hostname, src_asic_index, dst_asic_index)
+        )
+
+        backend_port_channels = extract_backend_portchannels(dut)
+        src_asic_next_hops, dst_asic_next_hops, src_prefix, dst_prefix = get_src_dst_asic_next_hops(
+            version,
+            dut,
+            src_asic,
+            dst_asic,
+            request,
+            backend_port_channels,
+        )
+
+        src_asic_router_mac = src_asic.get_router_mac()
+
+        yield {
+            "dut": dut,
+            "src_asic": src_asic,
+            "src_asic_index": src_asic_index,
+            "dst_asic": dst_asic,
+            "dst_asic_index": dst_asic_index,
+            "src_asic_next_hops": src_asic_next_hops,
+            "dst_asic_next_hops": dst_asic_next_hops,
+            "src_prefix": src_prefix,
+            "dst_prefix": dst_prefix,
+            "src_asic_router_mac": src_asic_router_mac,
+            "backend_port_channels": backend_port_channels,
+            "version": version,
+        }
 
     def test_bfd_traffic_remote_port_channel_shutdown(
         self,
@@ -43,10 +134,6 @@ class TestBfdTraffic(BfdBase):
             ("src", src_asic, src_prefix, src_asic_next_hops),
             ("dst", dst_asic, dst_prefix, dst_asic_next_hops),
         ]
-
-        with SafeThreadPoolExecutor(max_workers=8) as executor:
-            for _, asic, prefix, next_hops in src_dst_context:
-                executor.submit(create_and_verify_bfd_state, asic, prefix, dut, next_hops)
 
         dst_neighbor_ip = get_random_bgp_neighbor_ip_of_asic(dut, dst_asic_index, version)
         if not dst_neighbor_ip:
@@ -155,10 +242,6 @@ class TestBfdTraffic(BfdBase):
             ("dst", dst_asic, dst_prefix, dst_asic_next_hops),
         ]
 
-        with SafeThreadPoolExecutor(max_workers=8) as executor:
-            for _, asic, prefix, next_hops in src_dst_context:
-                executor.submit(create_and_verify_bfd_state, asic, prefix, dut, next_hops)
-
         dst_neighbor_ip = get_random_bgp_neighbor_ip_of_asic(dut, dst_asic_index, version)
         if not dst_neighbor_ip:
             pytest.skip("No BGP neighbor found on asic{} of dut {}".format(dst_asic_index, dut.hostname))
@@ -266,10 +349,6 @@ class TestBfdTraffic(BfdBase):
             ("dst", dst_asic, dst_prefix, dst_asic_next_hops),
         ]
 
-        with SafeThreadPoolExecutor(max_workers=8) as executor:
-            for _, asic, prefix, next_hops in src_dst_context:
-                executor.submit(create_and_verify_bfd_state, asic, prefix, dut, next_hops)
-
         dst_neighbor_ip = get_random_bgp_neighbor_ip_of_asic(dut, dst_asic_index, version)
         if not dst_neighbor_ip:
             pytest.skip("No BGP neighbor found on asic{} of dut {}".format(dst_asic_index, dut.hostname))
@@ -376,10 +455,6 @@ class TestBfdTraffic(BfdBase):
             ("src", src_asic, src_prefix, src_asic_next_hops),
             ("dst", dst_asic, dst_prefix, dst_asic_next_hops),
         ]
-
-        with SafeThreadPoolExecutor(max_workers=8) as executor:
-            for _, asic, prefix, next_hops in src_dst_context:
-                executor.submit(create_and_verify_bfd_state, asic, prefix, dut, next_hops)
 
         dst_neighbor_ip = get_random_bgp_neighbor_ip_of_asic(dut, dst_asic_index, version)
         if not dst_neighbor_ip:
